remove per-step serialization code
The serializer code ind the tests are specific to the data structures.
These will be cahnged to Pandas DataFrame objects soon.

This code is there only to support the task wrappers, which should go
anyway into a different repository.

Since the future "MVD3" should be an extensible format where attributes
may or may not exist, we could just use that for every single property
instead of having custom serializers for each one.

Change-Id: I1331cf70d01b10122eed7c9a7282e35afa799721