[NSETM-1170] A function to retrieve the region ids of layers in a brain region
This change extends the class LocalAtlas with a method that retrieves
the region identifers (e.g., the AIBS StructureIDs) of every layer specified in
a companion metadata file named metadata.json.

The function get_layer_volume lets you retrieve the volume where each voxel
is labeled with index of the layer it belongs to.

Currently, the use cases are the following:
* Retrieve the AIBS region identifiers of each layer in the mouse isocortex
* Retrieve the Paxinos-Watson region identifiers of each layer in the rat S1 brain region

Usage:
  from voxcell.nexus.voxelbrain import LocalAtlas
  import numpy.testing as npt
  atlas = LocalAtlas('atlases/aibs_isocortex_25_2017.nrrd')
  names, ids = atlas.get_layer_ids()
  layer_volume = atlas.get_layer_volume() # Returns a VoxelData object.
  npt.assert_array_equal(layer_volume.raw > 0, atlas.load_data('brain_regions.nrrd').raw > 0)

Change-Id: Iadd5640ae9f42cf71d26ebf5d412d948011372fe