Set the dtype of the default zero offset of VoxelData to np.float32 [NSETM-1513]
In the course of refactoring nse/atlas-buildling-tools [NSETM-1513],
I observed that the dtype of VoxelData's offset is float when no offset
vector is specified in the VoxelData constructor.

However, if the offset is specified in the constructor, then its dtype is
np.float32.

This change makes sure the dtype is np.float32 in every case.

Change-Id: I56d7c5dd2fea37b4a6230f6b80463489ad2d27da